Rename the library to pathlib and remove the word python from file and class names
Also rename the PythonPath class to NimPath and rename the pythonpathlib.nim and .nimble files to pathlib.nim and .nimble.

The reason for this change is that although the library tries to be as similar as possible to Python's pathlib, it is is a from scratch implementation which does not use Python at all. It is thus confusing in actual use to reference a library and classes with Python in their names.