Add the "touch" proc (similar to Python's)
Create a file at the given path.

If mode is given, it is applied to the file (otherwise full read/write access is
set)

If the file already exists, the function succeeds if exist_ok is true (which is
the default). In this case the file modification time will be updated to the
current time.

If the file already exists but exists_ok is set to false, an OSError exception
is raised.