Make / also work with Windows-like paths
Whether it uses / or \ as the path separator depends on the separator used on
the _first_ path. If no path separator is found on the first path, it is taken
from the second path. If no path separator is found on either of the paths the
default OS dir separator is used.